split implementation of SenderQueue for pipeline and non-pipeline
