+++ be/src/runtime/current_thread.h
@@ -4,6 +4,7 @@
 
 #include <string>
 
+#include "fmt/format.h"
 #include "gen_cpp/Types_types.h"
 #include "gutil/macros.h"
 #include "runtime/mem_tracker.h"
